% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/color_plane.R
\name{as_hexcol}
\alias{as_hexcol}
\alias{as_hexcol,RGBColorPlane-method}
\alias{as_hexcol,HexColorPlane-method}
\title{convert to hex colors}
\usage{
as_hexcol(x, ...)

\S4method{as_hexcol}{RGBColorPlane}(x)

\S4method{as_hexcol}{HexColorPlane}(x)
}
\arguments{
\item{x}{the object to convert}

\item{...}{extra args}
}
\value{
a character vector of ex colors
}
\description{
convert to hex colors
}
\seealso{
\link{rgb}
}
