% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/color_plane.R
\name{as_rgb}
\alias{as_rgb}
\alias{as_rgb,RGBColorPlane-method}
\alias{as_rgb,HexColorPlane-method}
\alias{as_rgb,ConstantColorPlane-method}
\title{convert to rgb colors}
\usage{
as_rgb(x, ...)

\S4method{as_rgb}{RGBColorPlane}(x)

\S4method{as_rgb}{HexColorPlane}(x)

\S4method{as_rgb}{ConstantColorPlane}(x)
}
\arguments{
\item{x}{the object to convert}

\item{...}{extra args}
}
\value{
a numeric matrix of rgb components
}
\description{
convert to rgb colors
}
\examples{
cp <- IntensityColorPlane(seq(1,100), cols=rainbow(25))
cl <- map_colors(cp, irange=c(0,50))
rgbcols <- as_rgb(cl)
}
