% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_plane.R
\name{col2hex}
\alias{col2hex}
\title{convert color name to hex character string}
\usage{
col2hex(cname, alpha = 1)
}
\arguments{
\item{cname}{one or more color names, e.g. "red"}

\item{alpha}{the value of the alpha channel, ranging from 0 to 1 (default is 1)}
}
\value{
a vector of hex color values, one per color name
}
\description{
convert color name to hex character string
}
