% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{ColorScale-class}
\alias{ColorScale-class}
\title{ColorScale}
\description{
ColorScale
}
\section{Slots}{

\describe{
\item{\code{irange}}{the intensity range of the scale}

\item{\code{threshold}}{the alpha thresholding range}

\item{\code{clr}}{a vector of hex colors}
}}

