% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/color_plane.R
\docType{class}
\name{ConstantColorPlane-class}
\alias{ConstantColorPlane-class}
\alias{ConstantColorPlane}
\title{ConstantColorPlane}
\usage{
ConstantColorPlane(clr)
}
\arguments{
\item{clr}{a single hex color as a `character` vector of length 1 defining the constant color.}
}
\value{
a new \code{\linkS4class{ConstantColorPlane}} instance
}
\description{
ConstantColorPlane constructor taking a single hex `character` vector defining a constant color plane.
}
\section{Slots}{

\describe{
\item{\code{clr}}{the constant color as hex value}
}}

\examples{

cp <- ConstantColorPlane(clr="#FF0000")
}
