% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_plane.R
\name{ConstantColorPlane}
\alias{ConstantColorPlane}
\title{ConstantColorPlane}
\usage{
ConstantColorPlane(clr)
}
\arguments{
\item{clr}{a single hex color as a `character` vector}
}
\description{
ConstantColorPlane constructor
}
