% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_plane.R
\name{DiscreteColorPlane}
\alias{DiscreteColorPlane}
\title{DiscreteColorPlane}
\usage{
DiscreteColorPlane(
  values,
  lookup = NULL,
  alpha = 1,
  cmap = rainbow(length(unique(values[!is.na(values)])))
)
}
\arguments{
\item{values}{the discrete values of the color map}

\item{lookup}{a "lookup table", which is a named list mapping discrete values to hex colors}

\item{alpha}{a value from 0 to 1 indicating transparency level}

\item{cmap}{a hex color vector}
}
\description{
DiscreteColorPlane constructor
}
