% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{get_color}
\alias{get_color}
\title{get_color}
\usage{
get_color(x, v, ...)
}
\arguments{
\item{x}{the color lookup table}

\item{v}{the intensity value(s)}

\item{...}{extra args}
}
\value{
a color value
}
\description{
get the color associated with one or more values
}
