% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/color_plane.R
\docType{class}
\name{HexColorPlane-class}
\alias{HexColorPlane-class}
\alias{HexColorPlane}
\title{HexColorPlane}
\usage{
HexColorPlane(clr)
}
\arguments{
\item{clr}{a field of colors}
}
\description{
HexColorPlane constructor
}
