% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/color_plane.R
\docType{class}
\name{IntensityColorPlane-class}
\alias{IntensityColorPlane-class}
\alias{IntensityColorPlane}
\title{IntensityColorPlane}
\usage{
IntensityColorPlane(intensity, cols = rainbow(255), alpha = 1)
}
\arguments{
\item{intensity}{a numeric vector of intensity values}

\item{cols}{a vector of hex character codes}

\item{alpha}{a vector of alpha values ranging from 0 to 1}
}
\value{
a new \code{\linkS4class{IntensityColorPlane}} instance
}
\description{
An association of intensities and colors

IntensityColorPlane constructor
}
\section{Slots}{

\describe{
\item{\code{intensity}}{a vector of intensity values}

\item{\code{alpha}}{a vector of alpha values}

\item{\code{colmap}}{a color map containing a vector of hex character codes}
}}

