% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_plane.R
\name{IntensityColorPlane}
\alias{IntensityColorPlane}
\title{IntensityColorPlane constructor}
\usage{
IntensityColorPlane(intensity, cols = rainbow(255), alpha = 1)
}
\description{
IntensityColorPlane constructor
}
