% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_plane.R
\name{multiply_alpha}
\alias{multiply_alpha}
\title{multiply rgb matrix with alpha channel}
\usage{
multiply_alpha(rgb, alpha)
}
\arguments{
\item{rgb}{matrix of colors}

\item{alpha}{channel}
}
\description{
multiply rgb matrix with alpha channel
}
\keyword{internal}
