% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_plane.R
\name{rgb2hex}
\alias{rgb2hex}
\title{convert rgb colors to hex colors}
\usage{
rgb2hex(r, g, b, alpha)
}
\arguments{
\item{r}{the red color component}

\item{g}{the green color component}

\item{b}{the blue color component}

\item{alpha}{the alpha component}
}
\value{
a hex color represenation as `character` vector
}
\description{
convert rgb colors to hex colors
}
