% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/color_plane.R
\docType{class}
\name{RGBColorPlane-class}
\alias{RGBColorPlane-class}
\alias{RGBColorPlane}
\title{RGBColorPlane}
\usage{
RGBColorPlane(clr)
}
\arguments{
\item{clr}{a field of colors}
}
\description{
RGBColorPlane constructor
}
