% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{ConstantColorPlane-class}
\alias{ConstantColorPlane-class}
\title{ConstantColorPlane}
\description{
ConstantColorPlane
}
\section{Slots}{

\describe{
\item{\code{clr}}{the constant color as hex value}
}}

\examples{

cp <- ConstantColorPlane(clr="#FF0000")
}
