% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{DiscreteColorPlane-class}
\alias{DiscreteColorPlane-class}
\title{DiscreteColorPlane}
\description{
DiscreteColorPlane
}
\section{Slots}{

\describe{
\item{\code{values}}{a vector of discrete integer values}

\item{\code{lookup}}{a lookup table mapping values to hex colors}

\item{\code{alpha}}{a vector of alpha values}
}}

